requirements.txt
requests
pandas
openpyxl # For Excel file support
python-dotenv